Intermediate commit. Here's what I have: RTC interrupt fires and the _rtcSec counter increments by one. Depending on the timing and bus synchronization it looks like the RTC->MODE1.COUNT.reg value could either be max val 32767 or over flowed. This can cause the delayRTCSteps method to be spoofed into thinking time is up when it really isnt. I have fixed this for delayRTCSteps however if we use millis() by polling this can still occur so I need to figure out how to make it so that millies() doesn't get spoofed.

+++ RTC.h
@@ -40,7 +40,7 @@ void             initRTC();
 void             disableRTC();
 volatile int64_t stepsRTC();
 volatile int64_t secondsRTC();
-void             delayRTCSteps( uint32_t steps );
+void             delayRTCSteps( int64_t steps );
 #ifdef __cplusplus
 }
 #endif
